relocate custom queue to a subfolder
1. It's not at the same level of abstraction as the client or
   transmissions.
2. It needs a different license notice to come along with it, so a
   separate folder helps distinguish it.
