+++ lib/libhoney/queueing.rb
@@ -1,0 +1,1 @@
+require 'libhoney/queueing/sized_queue_with_timeout'
