+++ lib/libhoney/queueing/sized_queue_with_timeout.rb
